Create PSARDingCode.txt
This is the code or "Think Script" you can copy and paste directly as a new study in ThinkorSwim for the Parabolic Stop and Reverse indicator. It adds sounds depending on which side of the candlesticks the study "flips" from.

This code creates an input called "alertOnFlip" that allows you to turn the alert function on or off. It then creates a variable called "flip" that is set to 1 if the ParabolicSAR indicator is above the current price and 0 if it is below. The "FlipSignal" plot simply displays the value of the "flip" variable.

The "alert()" function is then used to create the alert. It will trigger when the "FlipSignal" plot crosses above or below the value of 0. The "Alert.BAR" parameter specifies that the alert should only trigger once per bar, and the "Sound.Ding" parameter specifies that the alert should play a sound when triggered.